+++ metro.config.js
@@ -11,7 +11,7 @@ module.exports = (async () =>
         },
         resolver: {
             assetExts: assetExts.filter(ext => ext !== "svg"),
-            sourceExts: [...sourceExts, "svg"]
+            sourceExts: [...sourceExts, "jsx", "svg"]
         }
     };
 })();
